requirements.txt
Flask
flask_cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
PyJWT